Extract ExPlat implementation to :experimentation module
