+++ AutomatticTracks/build.gradle
@@ -6,20 +6,6 @@ plugins {
 
 repositories {
     google()
-    maven {
-        url 'https://a8c-libs.s3.amazonaws.com/android'
-        content {
-            includeGroup "org.wordpress"
-            includeGroup "org.wordpress.fluxc"
-            includeGroup "org.wordpress.wellsql"
-        }
-    }
-    maven {
-        url "https://a8c-libs.s3.amazonaws.com/android/jcenter-mirror"
-        content {
-            includeVersion "com.android.volley", "volley", "1.1.1"
-        }
-    }
 }
 
 dependencies {
@@ -34,16 +20,6 @@ dependencies {
     testImplementation 'org.mockito.kotlin:mockito-kotlin:2.2.11'
     testImplementation 'org.mockito:mockito-inline:3.8.0'
 
-    // Coroutines
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.2"
-
-    //FluxC
-    implementation "org.wordpress:fluxc:1.36.0"
-
-    //WordPress utils
-    implementation "org.wordpress:utils:2.3.0"
-
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.4.2"
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test:runner:1.3.0'
 }
