+++ experimentation/src/main/java/com/automattic/android/experimentation/example/ExampleExperiment.kt
@@ -1,0 +1,10 @@
+package com.automattic.android.experimentation.example
+
+import com.automattic.android.experimentation.ExPlat
+import com.automattic.android.experimentation.Experiment
+
+@Suppress("unused")
+class ExampleExperiment(exPlat: ExPlat) : Experiment(
+        name = "example_experiment",
+        exPlat = exPlat
+)
